webservice 修改
不太完善
我在publicationService写了一个getUserById但是这是应该在UserSerivce层的，我就没碰，自己写了一个。你可复制粘贴下

+++ src/service/PublicationService.java
@@ -4,31 +4,57 @@ import java.util.List;
 
 import javax.inject.Inject;
 
+import dao.FollowsDAO;
 import dao.PublicationDAO;
+import dao.UserDAO;
+import model.Follows;
 import model.Publication;
 import model.Response;
+import model.User;
 
 public class PublicationService {
 	@Inject
 	PublicationDAO publicationDAO;
+	@Inject
+	FollowsDAO followsDAO;
+	@Inject
+	UserDAO userDAO;
 
 	public void createNewPub(final Publication publication) {
 		publicationDAO.create(publication);
 	}
 	
-	public void deletePub(final Publication publication) {
-		publicationDAO.delete(publication);
+	public void deletePub(final Publication publication,User user) {
+		if (publication.getUser().getuId()==user.getuId()) {
+			publicationDAO.delete(publication);
+		}		
 	}
 	
-	public void updatePub(Publication publication, String imagePath, String comment, List<Response> responses) {
-		publicationDAO.update(publication, imagePath, comment, responses);
+	public void updatePub(Publication publication,User user, String imagePath, String comment, List<Response> responses) {
+		if (publication.getUser().getuId()==user.getuId()) {
+			publicationDAO.update(publication, imagePath, comment, responses);
+		}			
 	}
 	
-	public List<Publication> getAllPub() {
-		return publicationDAO.getAll();
+	
+	public List<Publication> getOtherAllPub(User user,User uFollew) {
+		List<Follows> getFollow = followsDAO.getFollow( uFollew.getuId(),user.getuId());
+		if (getFollow != null || getFollow.size() > 0) {
+			return publicationDAO.getAll(user);
+		}else 
+			return null;		
+	}
+	
+	public List<Publication> getOwnAllPub(User user){
+		return publicationDAO.getAll(user);
 	}
 	
 	public Publication getById(int pId) {
 		return publicationDAO.get(pId);
 	}
+	
+	public User getUserById(int uId) {
+		return userDAO.getById(uId);
+	}
+	
 }
